Implement method to normalize features
Previous experiements have shown that a neural network trained on boolean inputs does not work well when you start giving it probability values.
Some methods of imputation (notably gaussian based) imputes probabilies for booleans, so we use want to convert those back into booleans. We also take advantage of one hot to ensure that features with more options are not more likely to be all zero.