Add mutating datasets
The goal here is to be able to drop features without loading the entire dataset. For specific features its easy, just store the mask in the dataset.
For random features, we ensure each experiment starts with an identically seeded intiial random state, which ensures the same features are removed as long as iteration happens in the same order.
I could make it more stable by reseeding using the index, but that seems unneeded when I only iterate once in order anyways