Increase error debug steps
Been noticing with some architectures that eval and train mode have significantly different performance, so log the training error during validation steps. It is kinda slow compared to validation so maybe config to disable in the future
Also save the model at each step as we may wish to compare more than just the best
Finally, log initial validation error and use that instead of inf for initial comparison.