Allow passing in a text file containing newline seperated names for an image dataset
These text files already exist in CoPaint, so just reuse them for simplicity. Saves us the need to duplicate the entire dataset to a new folder just to do splits just to be able to use list directory to recreate that original file
To make this work, createCelebADataset now makes the folders for train/test/validate optional, and if unset it will default to using the root folder. There is also a new optional folder parameter that if set, will be the root for the split files