Work around softmax failure at extremely high values
Basically, if the model is overly confident, softmax will give you a probability of 1 with a variance of 0. That becomes 0/0, which breaks down
A probability of 0 with a variance of 0 would simililarly break
In either case, just set alpha values to infinity/0.
While changing this, also stopped calculating beta when its going to be unused (multiclass case)