Add option for decision makers to return the best class alongside best action
Want it for accuracy. In the paper, calling it "hard prediction" vs "soft prediction". Only makes sense when classes and labels have a 1 to 1 mapping
TODO: probably should define a version of best class that uses the loss function, though we aren't currently using that in the paper. Will be a lot easier once we swap loss function for a cost matrix